Upgrade focus restoration tests
